Refactor embedded steps for IDEA integration.
